requirements.txt
streamlit>=1.20.0
openai>=1.0.0 
anthropic>=0.20.0
pillow>=10.0.0
python-dotenv>=1.0.0
pandas>=1.3.0 
gradio
pytesseract>=0.3.10
surya-ocr==0.13.1
google-cloud-vision==3.4.2
# REMINDER: Setup Google Cloud authentication (e.g., gcloud auth application-default login)
qwen-agent[gui,rag,code_interpreter,mcp]